be more cautious of overriding AWS properties
only do so in DEV
