+++ common-lib/src/main/scala/com/gu/mediaservice/lib/config/StageIdentifier.scala
@@ -1,0 +1,17 @@
+package com.gu.mediaservice.lib.config
+
+import java.io.File
+
+import scala.io.Source.fromFile
+
+trait StageIdentifier {
+  final val stage: String = stageFromFile getOrElse "DEV"
+
+  val isProd: Boolean = stage == "PROD"
+  val isDev: Boolean = stage == "DEV"
+
+  private def stageFromFile: Option[String] = {
+    val file = new File("/etc/gu/stage")
+    if (file.exists) Some(fromFile(file).mkString.trim) else None
+  }
+}
